[#21324] mdd import script ported to 2.5 2.4x port by Ivo Grigull (loolarge)
from the tracker:

It should be working ok, but it does not have all options that the old script had. You can only specify the start frame
where to insert the animation, but not skip frames for have a different fps.

However, like the old script this will bloat the memory a bit, at some later point i would like to write a script that
only reads the current frame from the file when you move the timeline.