Fix #22905
[#22905] Adding objects with a grid size of <1m results in a 0 size object

The problem here is that the return value of the function
ED_object_new_primitive_matrix is a float, not an int.

This come from an old revision, before the two dia and depth
multiplication was using the function (two call to the same
function). This was change to only call the functio one time
and store the value in the "scale" variable, but the scale
was declare as int, not float.